{"schema":"decodex/commit/1","summary":"Split MCP autonomy proposal args","authority":"XY-1115"}
